+++ tests/integration/test_ai.py
@@ -372,6 +372,7 @@ def test_ai_spans_example_transaction(
 
     assert spans_consumer.get_spans(n=10) == [
         {
+            "_meta": matches_any(),
             "attributes": {
                 "gen_ai.conversation.id": {
                     "type": "string",
@@ -382,18 +383,35 @@ def test_ai_spans_example_transaction(
                     "value": matches_any(),
                 },
                 "gen_ai.context.window_size": {"type": "integer", "value": 128000},
+                "gen_ai.cost.cache_creation.input_tokens": {
+                    "type": "double",
+                    "value": 0.0,
+                },
+                "gen_ai.cost.cache_read.input_tokens": {
+                    "type": "double",
+                    "value": 0.0,
+                },
                 "gen_ai.cost.input_tokens": {"type": "double", "value": 2.45},
                 "gen_ai.cost.output_tokens": {"type": "double", "value": 1.3},
+                "gen_ai.cost.reasoning.output_tokens": {
+                    "type": "double",
+                    "value": 0.0,
+                },
                 "gen_ai.cost.total_tokens": {"type": "double", "value": 3.75},
                 "gen_ai.agent.name": {"type": "string", "value": "weather-chat"},
                 "gen_ai.function_id": {"type": "string", "value": "weather-chat"},
+                "gen_ai.input.messages": {
+                    "type": "string",
+                    "value": "Weather Prompt",
+                },
                 "gen_ai.operation.type": {"type": "string", "value": "agent"},
-                "gen_ai.input.messages": {"type": "string", "value": "Weather Prompt"},
+                "gen_ai.prompt": {"type": "string", "value": "Weather Prompt"},
                 "gen_ai.response.model": {"type": "string", "value": "gpt-4o"},
                 "gen_ai.output.messages": {
                     "type": "string",
                     "value": "True. \n\n- London: 61°F \n- San Francisco: 13°C",
                 },
+                "gen_ai.response.text": None,
                 "gen_ai.response.tokens_per_second": {"type": "double", "value": 130.0},
                 "gen_ai.usage.input_tokens": {"type": "integer", "value": 245},
                 "gen_ai.usage.output_tokens": {"type": "integer", "value": 65},
@@ -418,13 +436,14 @@ def test_ai_spans_example_transaction(
                 "sentry.op": {"type": "string", "value": "gen_ai.invoke_agent"},
                 "sentry.origin": {"type": "string", "value": "auto.vercelai.otel"},
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
                 "sentry.segment.name": {"type": "string", "value": "main"},
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",
@@ -481,8 +500,20 @@ def test_ai_spans_example_transaction(
                     "value": matches_any(),
                 },
                 "gen_ai.context.window_size": {"type": "integer", "value": 128000},
+                "gen_ai.cost.cache_creation.input_tokens": {
+                    "type": "double",
+                    "value": 0.0,
+                },
+                "gen_ai.cost.cache_read.input_tokens": {
+                    "type": "double",
+                    "value": 0.0,
+                },
                 "gen_ai.cost.input_tokens": {"type": "double", "value": 0.37},
                 "gen_ai.cost.output_tokens": {"type": "double", "value": 0.92},
+                "gen_ai.cost.reasoning.output_tokens": {
+                    "type": "double",
+                    "value": 0.0,
+                },
                 "gen_ai.cost.total_tokens": {"type": "double", "value": 1.29},
                 "gen_ai.agent.name": {"type": "string", "value": "weather-chat"},
                 "gen_ai.function_id": {"type": "string", "value": "weather-chat"},
@@ -495,6 +526,8 @@ def test_ai_spans_example_transaction(
                     "type": "string",
                     "value": "Another weather prompt",
                 },
+                "gen_ai.request.available_tools": None,
+                "gen_ai.request.messages": None,
                 "gen_ai.request.model": {"type": "string", "value": "gpt-4o"},
                 "gen_ai.response.finish_reasons": {
                     "type": "string",
@@ -509,6 +542,8 @@ def test_ai_spans_example_transaction(
                     "value": "gpt-4o-2024-08-06",
                 },
                 "gen_ai.response.tokens_per_second": {"type": "double", "value": 92.0},
+                "gen_ai.response.tool_calls": None,
+                "gen_ai.system": None,
                 "gen_ai.output.messages": {
                     "type": "string",
                     "value": "some_tool_calls",
@@ -537,13 +572,14 @@ def test_ai_spans_example_transaction(
                 "sentry.op": {"type": "string", "value": "gen_ai.generate_text"},
                 "sentry.origin": {"type": "string", "value": "auto.vercelai.otel"},
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
                 "sentry.segment.name": {"type": "string", "value": "main"},
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",
@@ -620,6 +656,8 @@ def test_ai_spans_example_transaction(
                 "network.peer.address": {"type": "string", "value": "162.159.140.245"},
                 "network.peer.port": {"type": "integer", "value": 443},
                 "otel.kind": {"type": "string", "value": "CLIENT"},
+                "sentry.action": {"type": "string", "value": "POST"},
+                "sentry.kind": {"type": "string", "value": "CLIENT"},
                 "sentry.category": {"type": "string", "value": "http"},
                 "sentry.description": {
                     "type": "string",
@@ -645,6 +683,8 @@ def test_ai_spans_example_transaction(
                     "value": "auto.http.otel.node_fetch",
                 },
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
@@ -652,7 +692,6 @@ def test_ai_spans_example_transaction(
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.status_code": {"type": "string", "value": "200"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",
@@ -693,6 +732,7 @@ def test_ai_spans_example_transaction(
             "trace_id": "a9351cd574f092f6acad48e250981f11",
         },
         {
+            "_meta": matches_any(),
             "attributes": {
                 "gen_ai.operation.type": {"type": "string", "value": "tool"},
                 "gen_ai.tool.call.id": {
@@ -712,6 +752,8 @@ def test_ai_spans_example_transaction(
                     '(56°F)","temperatureC":13,"temperatureF":56,"condition":"Haze","humidity":"88%","windSpeed":"4 '
                     'km/h"}',
                 },
+                "gen_ai.tool.input": None,
+                "gen_ai.tool.output": None,
                 "gen_ai.tool.type": {"type": "string", "value": "function"},
                 "operation.name": {
                     "type": "string",
@@ -733,13 +775,14 @@ def test_ai_spans_example_transaction(
                 "sentry.op": {"type": "string", "value": "gen_ai.execute_tool"},
                 "sentry.origin": {"type": "string", "value": "auto.vercelai.otel"},
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
                 "sentry.segment.name": {"type": "string", "value": "main"},
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",
@@ -779,6 +822,8 @@ def test_ai_spans_example_transaction(
                 "network.peer.address": {"type": "string", "value": "5.9.243.187"},
                 "network.peer.port": {"type": "integer", "value": 443},
                 "otel.kind": {"type": "string", "value": "CLIENT"},
+                "sentry.action": {"type": "string", "value": "GET"},
+                "sentry.kind": {"type": "string", "value": "CLIENT"},
                 "sentry.category": {"type": "string", "value": "http"},
                 "sentry.description": {
                     "type": "string",
@@ -804,6 +849,8 @@ def test_ai_spans_example_transaction(
                     "value": "auto.http.otel.node_fetch",
                 },
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
@@ -811,7 +858,6 @@ def test_ai_spans_example_transaction(
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.status_code": {"type": "string", "value": "200"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",
@@ -845,6 +891,7 @@ def test_ai_spans_example_transaction(
             "trace_id": "a9351cd574f092f6acad48e250981f11",
         },
         {
+            "_meta": matches_any(),
             "attributes": {
                 "gen_ai.operation.type": {"type": "string", "value": "tool"},
                 "gen_ai.tool.call.id": {
@@ -864,6 +911,8 @@ def test_ai_spans_example_transaction(
                     'cloudy","humidity":"72%","windSpeed":"21 '
                     'km/h"}',
                 },
+                "gen_ai.tool.input": None,
+                "gen_ai.tool.output": None,
                 "gen_ai.tool.type": {"type": "string", "value": "function"},
                 "operation.name": {
                     "type": "string",
@@ -885,13 +934,14 @@ def test_ai_spans_example_transaction(
                 "sentry.op": {"type": "string", "value": "gen_ai.execute_tool"},
                 "sentry.origin": {"type": "string", "value": "auto.vercelai.otel"},
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
                 "sentry.segment.name": {"type": "string", "value": "main"},
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",
@@ -931,6 +981,8 @@ def test_ai_spans_example_transaction(
                 "network.peer.address": {"type": "string", "value": "5.9.243.187"},
                 "network.peer.port": {"type": "integer", "value": 443},
                 "otel.kind": {"type": "string", "value": "CLIENT"},
+                "sentry.action": {"type": "string", "value": "GET"},
+                "sentry.kind": {"type": "string", "value": "CLIENT"},
                 "sentry.category": {"type": "string", "value": "http"},
                 "sentry.description": {
                     "type": "string",
@@ -956,6 +1008,8 @@ def test_ai_spans_example_transaction(
                     "value": "auto.http.otel.node_fetch",
                 },
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
@@ -963,7 +1017,6 @@ def test_ai_spans_example_transaction(
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.status_code": {"type": "string", "value": "200"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",
@@ -1008,8 +1061,20 @@ def test_ai_spans_example_transaction(
                     "value": matches_any(),
                 },
                 "gen_ai.context.window_size": {"type": "integer", "value": 128000},
+                "gen_ai.cost.cache_creation.input_tokens": {
+                    "type": "double",
+                    "value": 0.0,
+                },
+                "gen_ai.cost.cache_read.input_tokens": {
+                    "type": "double",
+                    "value": 0.0,
+                },
                 "gen_ai.cost.input_tokens": {"type": "double", "value": 2.08},
                 "gen_ai.cost.output_tokens": {"type": "double", "value": 0.38},
+                "gen_ai.cost.reasoning.output_tokens": {
+                    "type": "double",
+                    "value": 0.0,
+                },
                 "gen_ai.cost.total_tokens": {"type": "double", "value": 2.46},
                 "gen_ai.agent.name": {"type": "string", "value": "weather-chat"},
                 "gen_ai.function_id": {"type": "string", "value": "weather-chat"},
@@ -1022,6 +1087,8 @@ def test_ai_spans_example_transaction(
                     "type": "string",
                     "value": "Some AI Prompt about " "the Wheather",
                 },
+                "gen_ai.request.available_tools": None,
+                "gen_ai.request.messages": None,
                 "gen_ai.request.model": {"type": "string", "value": "gpt-4o"},
                 "gen_ai.response.finish_reasons": {
                     "type": "string",
@@ -1042,7 +1109,9 @@ def test_ai_spans_example_transaction(
                     "- London: 61°F \n"
                     "- San Francisco: 13°C",
                 },
+                "gen_ai.response.text": None,
                 "gen_ai.response.tokens_per_second": {"type": "double", "value": 38.0},
+                "gen_ai.system": None,
                 "gen_ai.provider.name": {"type": "string", "value": "openai.responses"},
                 "gen_ai.usage.input_tokens": {"type": "integer", "value": 208},
                 "gen_ai.usage.output_tokens": {"type": "integer", "value": 19},
@@ -1067,13 +1136,14 @@ def test_ai_spans_example_transaction(
                 "sentry.op": {"type": "string", "value": "gen_ai.generate_text"},
                 "sentry.origin": {"type": "string", "value": "auto.vercelai.otel"},
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
                 "sentry.segment.name": {"type": "string", "value": "main"},
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",
@@ -1147,6 +1217,8 @@ def test_ai_spans_example_transaction(
                 "network.peer.address": {"type": "string", "value": "162.159.140.245"},
                 "network.peer.port": {"type": "integer", "value": 443},
                 "otel.kind": {"type": "string", "value": "CLIENT"},
+                "sentry.action": {"type": "string", "value": "POST"},
+                "sentry.kind": {"type": "string", "value": "CLIENT"},
                 "sentry.category": {"type": "string", "value": "http"},
                 "sentry.description": {
                     "type": "string",
@@ -1172,6 +1244,8 @@ def test_ai_spans_example_transaction(
                     "value": "auto.http.otel.node_fetch",
                 },
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
@@ -1179,7 +1253,6 @@ def test_ai_spans_example_transaction(
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.status_code": {"type": "string", "value": "200"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",
@@ -1242,15 +1315,17 @@ def test_ai_spans_example_transaction(
                 "sentry.op": {"type": "string", "value": "gen_ai.invoke_agent"},
                 "sentry.origin": {"type": "string", "value": "manual"},
                 "sentry.platform": {"type": "string", "value": "node"},
+                "sentry.relay.ingress": {"type": "string", "value": "legacy"},
+                "sentry.relay.pipeline": {"type": "string", "value": "transaction"},
                 "sentry.sample_rate": {"type": "integer", "value": 1},
                 "sentry.sdk.name": {"type": "string", "value": "raven-node"},
                 "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
                 "sentry.segment.id": {"type": "string", "value": "657cf984a6a4e59b"},
                 "sentry.segment.name": {"type": "string", "value": "main"},
                 "sentry.source": {"type": "string", "value": "custom"},
+                "sentry.span.source": {"type": "string", "value": "custom"},
                 "sentry.status": {"type": "string", "value": "ok"},
                 "sentry.trace.status": {"type": "string", "value": "ok"},
-                "sentry.transaction": {"type": "string", "value": "main"},
                 "sentry.transaction.op": {
                     "type": "string",
                     "value": "gen_ai.invoke_agent",